Submit file made for single AJAX request
