+++ mobile_application/templates/index.html
@@ -34,6 +34,7 @@
 	<script src = '\static\js\mouth.js' type = 'module'></script>
 	<script src = '\static\js\nose.js' type = 'module'></script>
 	<script src = '\static\js\brows.js' type = 'module'></script>
+	<script src = '\static\js\submit.js' type = 'module'></script>
 </head>
 
 <body>
